moved mapbox component to shared, and integrated into view-carpool-details page
