+++ src/app/_shared/components/mapbox/mapbox.component.css
@@ -1,0 +1,9 @@
+#map {
+  position: absolute;
+  width: 100vw;
+  height: 40vh;
+
+  /* extending background to overcome the global padding at app.component.css level */
+  margin-left: -10px;
+  margin-right: -10px;
+}
