+++ src/app/_shared/components/mapbox/mapbox.component.html
@@ -1,0 +1,1 @@
+<div id="map"></div>
