+++ src/app/_shared/services/api-database.service.ts
@@ -25,6 +25,10 @@ export class ApiDatabaseService {
       .valueChanges({idField: 'carpoolId'});
   }
 
+  showCarpool(carpoolId: string) {
+    return this.db.collection('carpools').doc<ICarpool>(carpoolId).valueChanges();
+  }
+
   addUserToCarpool(carpoolId: string, userId: string) {
     // console.log(`updating ${carpoolId} with ${userId}`);
     return this.showUser(userId).pipe(
