--- src/app/mapbox/mapbox.component.css
@@ -1,6 +1,0 @@
-#map {
-    position: absolute;
-    top: 100px;
-    width: 100%;
-    height: 60%;
-}
