--- src/app/mapbox/mapbox.component.html
@@ -1,8 +1,0 @@
-<input type="text" [(ngModel)]="message" placeholder="your message...">
-<!-- <h1>Markers</h1>
-<div *ngFor="let marker of markers | async">
-  <button (click)="flyTo(marker)">{{ marker.properties.message }}</button>
-  <button (click)="removeMarker(marker)">Delete</button>
-</div> -->
-
-<div class="map" id="map"></div>
